# Improve message list diff performance