--- a/stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/adapter/MessageListItem.kt
+++ b/stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/adapter/MessageListItem.kt
@@ -9,8 +9,6 @@ import io.getstream.chat.android.client.models.ChannelUserRead
 import io.getstream.chat.android.client.models.Message
 import io.getstream.chat.android.client.models.User
 import java.util.Date
-import java.util.zip.CRC32
-import java.util.zip.Checksum
 
 /**
  * [MessageListItem] represents elements that are displayed in a [com.getstream.sdk.chat.view.MessageListView].
@@ -24,16 +22,13 @@ import java.util.zip.Checksum
 public sealed class MessageListItem {
 
     public fun getStableId(): Long {
-        val checksum: Checksum = CRC32()
-        val plaintext = when (this) {
-            is TypingItem -> "id_typing"
-            is ThreadSeparatorItem -> "id_thread_separator"
-            is MessageItem -> message.id
-            is DateSeparatorItem -> date.toString()
-            is LoadingMoreIndicatorItem -> "id_loading_more_indicator"
+        return when (this) {
+            is TypingItem -> TYPING_ITEM_STABLE_ID
+            is ThreadSeparatorItem -> THREAD_SEPARATOR_ITEM_STABLE_ID
+            is MessageItem -> message.id.hashCode().toLong()
+            is DateSeparatorItem -> date.time
+            is LoadingMoreIndicatorItem -> LOADING_MORE_INDICATOR_STABLE_ID
         }
-        checksum.update(plaintext.toByteArray(), 0, plaintext.toByteArray().size)
-        return checksum.value
     }
 
     public data class DateSeparatorItem(
@@ -44,7 +39,7 @@ public sealed class MessageListItem {
         val message: Message,
         val positions: List<Position> = listOf(),
         val isMine: Boolean = false,
-        val messageReadBy: List<ChannelUserRead> = listOf()
+        val messageReadBy: List<ChannelUserRead> = listOf(),
     ) : MessageListItem() {
         public val isTheirs: Boolean
             get() = !isMine
@@ -65,5 +60,11 @@ public sealed class MessageListItem {
         TOP,
         MIDDLE,
         BOTTOM,
+    }
+
+    private companion object {
+        const val TYPING_ITEM_STABLE_ID = 1L
+        const val THREAD_SEPARATOR_ITEM_STABLE_ID = 2L
+        const val LOADING_MORE_INDICATOR_STABLE_ID = 3L
     }
 }