### Description

This PR improves the performance of `MessageListItemDiffCallback`

What was measured:
The time span between `submitList` invocation and `commitCallback`. (Provided that all the messages in the "general" channel are loaded up to the 6th of December and we post a new message to the channel). 

Results vary greatly from time to time but on average 1 diff took:
Before the change - 300ms (150ms - 800ms)
After the change - 4ms (1ms - 8ms)

### Checklist

- [X] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [X] PR targets the `develop` branch
- [ ] Changelog updated with client-facing changes
- [ ] New code is covered by unit tests
- [ ] Comparison screenshots added for visual changes
- [X] Reviewers added